Refs #17. moved scaffolding from singlextal to mcvine.workflow.sampleassembly
