+++ tests/mcvine_workflow/sampleassembly/scaffolding/DGSresolution_TestCase.py
