+++ tests/mcvine_workflow/sampleassembly/scaffolding/KVO-rotated.yml
