+++ tests/mcvine_workflow/sampleassembly/scaffolding/KVO.yml
