+++ tests/mcvine_workflow/sampleassembly/scaffolding/expected-sampleassembly/KVO-scatterer.xml
