+++ tests/mcvine_workflow/sampleassembly/scaffolding/expected-sampleassembly/sampleassembly.xml
