+++ tests/mcvine_workflow/sampleassembly/scaffolding/expected-sampleassembly2/sampleassembly.xml
