+++ tests/mcvine_workflow/sampleassembly/scaffolding/expected-scatterer-DGSResolution_kernel/sample.DGSResolution.xyz
