+++ tests/mcvine_workflow/sampleassembly/scaffolding/expected-scatterer-deltafunction_kernel/sample.deltafunction-scatterer.xml
