+++ tests/mcvine_workflow/sampleassembly/scaffolding/expected-scatterer-phonon_kernels/sample.phonon.xyz
