+++ tests/mcvine_workflow/sampleassembly/scaffolding/expected-scatterer-spinwave_kernel/sample.spinwave.xyz
