+++ tests/mcvine_workflow/sampleassembly/scaffolding/sampleassembly_TestCase.py
